9/5 6:37am Fades Now Working Correctly
