+++ Assets/Scenes/LevelSelect.unity
@@ -469,6 +469,7 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 889869936}
+  - {fileID: 2020761020}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -507,7 +508,6 @@ RectTransform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 2020761020}
   - {fileID: 1561548042}
   - {fileID: 1655437840}
   m_Father: {fileID: 516017441}
@@ -793,7 +793,7 @@ RectTransform:
   m_Children:
   - {fileID: 384406686}
   m_Father: {fileID: 889869936}
-  m_RootOrder: 1
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
@@ -813,6 +813,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   thisFadePanel: {fileID: 2020761025}
+  sceneToLoad: TestLevel
 --- !u!114 &1561548044
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -857,21 +858,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2020761026}
-        m_TargetAssemblyTypeName: Fade, Assembly-CSharp
-        m_MethodName: FadeOutFunc
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
       - m_Target: {fileID: 1561548043}
         m_TargetAssemblyTypeName: LevelSelect, Assembly-CSharp
-        m_MethodName: GoToEarth
+        m_MethodName: StartFade
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -931,6 +920,7 @@ GameObject:
   - component: {fileID: 1655437843}
   - component: {fileID: 1655437842}
   - component: {fileID: 1655437841}
+  - component: {fileID: 1655437844}
   m_Layer: 5
   m_Name: ExitToTitle
   m_TagString: Untagged
@@ -952,7 +942,7 @@ RectTransform:
   m_Children:
   - {fileID: 1395911991}
   m_Father: {fileID: 889869936}
-  m_RootOrder: 2
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
@@ -1003,21 +993,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2020761026}
-        m_TargetAssemblyTypeName: Fade, Assembly-CSharp
-        m_MethodName: FadeOutFunc
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-      - m_Target: {fileID: 1561548043}
+      - m_Target: {fileID: 1655437844}
         m_TargetAssemblyTypeName: LevelSelect, Assembly-CSharp
-        m_MethodName: ExitToTitle
+        m_MethodName: StartFade
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -1065,6 +1043,20 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1655437839}
   m_CullTransparentMesh: 1
+--- !u!114 &1655437844
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1655437839}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0f01b126090aa49879a118903beadfd9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  thisFadePanel: {fileID: 2020761025}
+  sceneToLoad: TitleScreen
 --- !u!1 &2020761019
 GameObject:
   m_ObjectHideFlags: 0
@@ -1097,8 +1089,8 @@ RectTransform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
-  m_Father: {fileID: 889869936}
-  m_RootOrder: 0
+  m_Father: {fileID: 516017441}
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 1, y: 1}
