+++ Assets/Scripts/IngameUI/PausePanel.cs
@@ -27,17 +27,8 @@ public class PausePanel : MonoBehaviour {
             Time.timeScale = 1f;
         }
     }
-
     public void Resume() {
         pausePanel.SetActive(false);
         Time.timeScale = 1f;
-    }
-
-    public void GiveUp() {
-        SceneManager.LoadScene("LevelSelect");
-    }
-
-    public void ExitGame() {
-        Application.Quit();
     }
 }
