+++ Assets/Scripts/WeaponScripts/MutagenicNanobots.cs
@@ -4,7 +4,6 @@ public class MutagenicNanobots : Weapon {
     public NanoManager nanoManager;
     private GameObject _playerPosition;
     private Vector3 _center;
-    public GameObject nanoBot;
     public float spawnDelay;
 
     void Start() {
@@ -22,7 +21,6 @@ public class MutagenicNanobots : Weapon {
         if (nanoManager.currentNanoBots < nanoManager.maxNanoBots) {
             Vector3 offset = new Vector3(_center.x, _center.y + 3, 0);
             if (nanoManager.currentNanoBots < nanoManager.maxNanoBots) {
-                //Instantiate(nanoBot, offset, Quaternion.identity);
                 GameObject nano = PoolManager.pm.nanoPool.GetPooledGameObject();
                 nano.SetActive(true);
                 nano.transform.position = offset;
