--- Assets/Visuals/Epic Toon FX/Upgrade/ETFX 2019.4.24f1 URP Upgrade.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f613353c655d1ab439166e4070021529
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
