--- Assets/ZengsStash/Feel/FeelDemos/FeelTemplatesDemo.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6122a0a4380825642a9ed4147100f1f9
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
