--- Assets/ZengsStash/Feel/NiceVibrations/OlderVersions/v1.7/NiceVibrations-v-1-7.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b91fdb0369220fc47808e4469a803a6f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
