+++ ProjectSettings/TagManager.asset
@@ -81,3 +81,6 @@ TagManager:
   - name: UI
     uniqueID: 1906808429
     locked: 0
+  - name: FadePanel
+    uniqueID: 3818868637
+    locked: 0
